// rne-start/src/types/IQuestDetails.ts
export interface IQuestDetails {
  id: string;
  apiPath: string;
  name: string;
  mainGenre: string;
  location: string;
  banner: string;
  levelOfFear: string;
  difficultyLevel: string;
  contactLevel: string;
  teamSize: string;
  ageFrom: string;
  duration: string;
  interactive: boolean;
  description: string[];
  additionalDescription: string[];
  modes: string[];
}